third_party/WebKit/LayoutTests/inspector/console/console-revoke-error-in-worker.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

var worker;

function createPromise()
{
    worker = new Worker("resources/worker-with-defer-handled-promise.js");
}

function handlePromiseRejection()
{
    worker.postMessage("");
}

function test() {
  ConsoleModel.consoleModel.addEventListener(
      ConsoleModel.ConsoleModel.Events.MessageAdded, ConsoleTestRunner.wrapListener(messageAdded));
  ConsoleModel.consoleModel.addEventListener(
      ConsoleModel.ConsoleModel.Events.MessageUpdated, ConsoleTestRunner.wrapListener(messageUpdated));

  TestRunner.addResult('Creating worker with promise');
  TestRunner.evaluateInPageWithTimeout('createPromise()');

  function messageAdded(event) {
    TestRunner.addResult('');
    TestRunner.addResult('Message added: ' + event.data.level + ' ' + event.data.type);

    if (event.data.level === ConsoleModel.ConsoleMessage.MessageLevel.Error) {
      ConsoleTestRunner.dumpConsoleCounters();
      TestRunner.addResult('');
      TestRunner.addResult('Handling promise');
      TestRunner.evaluateInPageWithTimeout('handlePromiseRejection()');
    }
  }

  function messageUpdated() {
    ConsoleTestRunner.dumpConsoleCounters();
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>Tests that console revokes lazily handled promise rejections.</p>

</body>
</html>
